requirements.txt
bcrypt
pyjwt
pytest
Flask-WTF
